[provider.cube] refs #10

+++ dune/stuff/grid/provider/cube.hh
@@ -222,7 +222,7 @@ private:
       DUNE_THROW(Exceptions::wrong_input_given,
                  "Given vector is too short: should be " << dimDomain << ", is " << in.size() << "!");
     std::array<unsigned int, dimDomain> ret;
-    for (unsigned int ii = 0; ii < dimDomain; ++ii)
+    for (size_t ii = 0; ii < dimDomain; ++ii)
       ret[ii] = in[ii];
     return ret;
   } // ... parse_array(...)
@@ -232,7 +232,7 @@ private:
                                                const size_t num_refinements)
   {
     static_assert(variant == 1 || variant == 2, "variant has to be 1 or 2!");
-    for (unsigned int dd = 0; dd < dimDomain; ++dd) {
+    for (size_t dd = 0; dd < dimDomain; ++dd) {
       if (!(lower_left[dd] < upper_right[dd]))
         DUNE_THROW(Exceptions::wrong_input_given,
                    "lower_left has to be elementwise smaller than upper_right!\n\n" << lower_left[dd] << " vs. "
